pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "github-tech-doc-generator"
version = "0.1.0"
description = "Backend-first MVP for generating technical docs from GitHub repositories."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pydantic-settings>=2.3.0,<3.0.0",
  "httpx>=0.27.0,<0.28.0",
  "PyJWT[crypto]>=2.8.0,<3.0.0",
  "PyYAML>=6.0.1,<7.0.0",
  "redis>=5.0.0,<6.0.0",
  "arq>=0.26.0,<0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0.0",
  "pytest-asyncio>=0.23.0,<0.24.0",
  "fakeredis>=2.23.0,<3.0.0",
]

[tool.pytest.ini_options]
addopts = "-q -p no:cacheprovider -p no:tmpdir"
testpaths = ["tests"]
norecursedirs = [".tmp", ".pytest-tmp", "tmpbase", "pytest-tmp"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
